requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.7
openai>=1.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.0
tiktoken>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi
uvicorn
python-multipart
jinja2
starlette
docling[all]
pypdf
pandas
openpyxl
python-docx
